docs/benchmarks/async_benchmark.py
#!/usr/bin/env python3
# SPDX-License-Identifier: MIT
"""Benchmark script to demonstrate async optimization improvements.

This script simulates concurrent operations to measure the performance
benefits of the async refactoring. It measures:
- Concurrent image preparations (CPU-bound PIL operations)
- Concurrent file I/O operations
- Mixed concurrent workloads

Run with: python docs/benchmarks/async_benchmark.py
"""

import asyncio
import base64
import pathlib
import tempfile
import time
from typing import Any

from PIL import Image


async def benchmark_concurrent_image_prep(num_operations: int = 10) -> dict[str, Any]:
    """Benchmark concurrent image resize operations.

    This simulates the prepare_reference_image() function running PIL operations
    in thread pools, allowing multiple resize operations to run concurrently.
    """

    async def resize_image_task(task_id: int) -> float:
        """Simulate a CPU-intensive image resize operation."""
        start = time.perf_counter()

        def _do_resize():
            # Create a test image
            img = Image.new("RGB", (2048, 2048), color=(task_id * 20, 100, 150))
            # Simulate resize operations
            img = img.resize((1280, 720), Image.Resampling.LANCZOS)
            # Simulate crop
            img = img.crop((0, 0, 1280, 720))
            return img.size

        # Run in thread pool (simulating anyio.to_thread.run_sync)
        await asyncio.to_thread(_do_resize)

        return time.perf_counter() - start

    start_time = time.perf_counter()
    task_times = await asyncio.gather(*[resize_image_task(i) for i in range(num_operations)])
    total_time = time.perf_counter() - start_time

    return {
        "test": "Concurrent Image Preparations",
        "operations": num_operations,
        "total_time": total_time,
        "avg_time_per_op": sum(task_times) / len(task_times),
        "speedup_vs_sequential": sum(task_times) / total_time,
    }


async def benchmark_concurrent_file_writes(num_operations: int = 20) -> dict[str, Any]:
    """Benchmark concurrent async file write operations.

    This simulates async file writes using aiofiles, showing improved
    throughput under concurrent I/O load.
    """

    async def write_file_task(tmp_dir: pathlib.Path, task_id: int) -> float:
        """Simulate an async file write operation."""
        start = time.perf_counter()

        # Simulate writing a file (using standard asyncio to_thread for portability)
        file_path = tmp_dir / f"test_file_{task_id}.bin"
        data = b"x" * (1024 * 100)  # 100KB

        def _write():
            with open(file_path, "wb") as f:
                f.write(data)

        await asyncio.to_thread(_write)

        return time.perf_counter() - start

    with tempfile.TemporaryDirectory() as tmp_dir:
        tmp_path = pathlib.Path(tmp_dir)

        start_time = time.perf_counter()
        task_times = await asyncio.gather(*[write_file_task(tmp_path, i) for i in range(num_operations)])
        total_time = time.perf_counter() - start_time

    return {
        "test": "Concurrent File Writes",
        "operations": num_operations,
        "total_time": total_time,
        "avg_time_per_op": sum(task_times) / len(task_times),
        "speedup_vs_sequential": sum(task_times) / total_time,
    }


async def benchmark_base64_decode(num_operations: int = 15) -> dict[str, Any]:
    """Benchmark concurrent base64 decoding operations.

    This simulates the download_image() function decoding large base64
    strings in thread pools.
    """

    async def decode_task(task_id: int) -> float:
        """Simulate CPU-intensive base64 decoding."""
        start = time.perf_counter()

        # Create large base64 string (simulating ~1MB image)
        data = b"x" * (1024 * 1024)
        encoded = base64.b64encode(data).decode("utf-8")

        # Decode in thread pool
        await asyncio.to_thread(base64.b64decode, encoded)

        return time.perf_counter() - start

    start_time = time.perf_counter()
    task_times = await asyncio.gather(*[decode_task(i) for i in range(num_operations)])
    total_time = time.perf_counter() - start_time

    return {
        "test": "Concurrent Base64 Decoding",
        "operations": num_operations,
        "total_time": total_time,
        "avg_time_per_op": sum(task_times) / len(task_times),
        "speedup_vs_sequential": sum(task_times) / total_time,
    }


async def benchmark_mixed_workload() -> dict[str, Any]:
    """Benchmark mixed concurrent operations.

    This simulates a realistic scenario with multiple types of operations
    running concurrently: image processing, file I/O, and data encoding.
    """
    start_time = time.perf_counter()

    # Run different types of operations concurrently
    results = await asyncio.gather(
        benchmark_concurrent_image_prep(num_operations=5),
        benchmark_concurrent_file_writes(num_operations=10),
        benchmark_base64_decode(num_operations=8),
    )

    total_time = time.perf_counter() - start_time
    total_ops = sum(r["operations"] for r in results)

    return {
        "test": "Mixed Concurrent Workload",
        "operations": total_ops,
        "total_time": total_time,
        "subtasks": [r["test"] for r in results],
    }


def print_results(result: dict[str, Any]) -> None:
    """Pretty print benchmark results."""
    print(f"\n{'=' * 70}")
    print(f"TEST: {result['test']}")
    print(f"{'=' * 70}")
    print(f"  Operations:        {result['operations']}")
    print(f"  Total Time:        {result['total_time']:.3f}s")

    if "avg_time_per_op" in result:
        print(f"  Avg Time/Op:       {result['avg_time_per_op']:.4f}s")
        print(f"  Speedup Factor:    {result['speedup_vs_sequential']:.2f}x")
        print("    (vs sequential execution)")

    if "subtasks" in result:
        print(f"  Subtasks:          {', '.join(result['subtasks'])}")


async def main():
    """Run all benchmarks."""
    print("\n" + "=" * 70)
    print("ASYNC OPTIMIZATION BENCHMARK")
    print("=" * 70)
    print("\nDemonstrating performance improvements from:")
    print("  • anyio thread pools for CPU-bound operations")
    print("  • aiofiles for non-blocking file I/O")
    print("  • Streaming downloads with async iteration")
    print("\nNOTE: Speedup factors show theoretical max speedup if operations")
    print("      were run sequentially. Higher = better concurrency.")

    # Run individual benchmarks
    image_results = await benchmark_concurrent_image_prep(num_operations=10)
    print_results(image_results)

    file_results = await benchmark_concurrent_file_writes(num_operations=20)
    print_results(file_results)

    decode_results = await benchmark_base64_decode(num_operations=15)
    print_results(decode_results)

    # Run mixed workload
    mixed_results = await benchmark_mixed_workload()
    print_results(mixed_results)

    print("\n" + "=" * 70)
    print("SUMMARY")
    print("=" * 70)
    print("\nThe async optimizations enable true concurrency for:")
    print("  1. CPU-bound operations (PIL, base64) via thread pools")
    print("  2. I/O-bound operations (file reads/writes) via async I/O")
    print("  3. Mixed workloads running simultaneously")
    print("\nWith Python 3.14's free-threading, these gains will be even larger!")
    print("=" * 70 + "\n")


if __name__ == "__main__":
    asyncio.run(main())
